fixed db migrations

+++ SivenaBackend/database/migrations/2024_06_06_065955_create_posts_table.php
@@ -22,7 +22,7 @@ class CreatePostsTable extends Migration
 
             // Foreign Keys
             $table->unsignedBigInteger('category_id'); // Reference to categories
-            $table->unsignedBigInteger('author_id'); // Reference to users (author)
+            $table->unsignedInteger('author_id'); // Reference to users (author)
 
             // Timestamps
             $table->timestamps();
